+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"09.07.2026 13:44"
-#define NE_VERSION   L"2026.07.09.13"
+#define NE_PUBLISHED L"09.07.2026 17:23"
+#define NE_VERSION   L"2026.07.09.17"